dataset/extract/requirements.txt
pydantic<2.0.0
jsbeautifier
pandas
PyGithub
GitPython
tqdm
cchardet
blake3
esprima
javalang
pyjsparser
